Renamed logging → logging_utils to avoid absolute_import.
This should partially address #377, as absolute_import seems to work
differently on Windows for some reason?
