+++ qutip/control/tslotcomp.py
@@ -71,7 +71,7 @@ import timeit
 # QuTiP control modules
 import qutip.control.errors as errors
 # QuTiP logging
-import qutip.logging as logging
+import qutip.logging_utils as logging
 logger = logging.get_logger()
