+++ qutip/logging_utils.py
